MB-33799 Return nil snaphsots if mainstore recovery points are not initialized
When persisting snapshots, it is possible that backstore recovery points
are created while mainstore recovery points are not created. In such cases,
a call to GetSnapshots() will panic and indexer will not be able to come
out of the panic loop.

In this patch, we attempt to fix the problem by returning nil snapshots when
mainstore recovery points are not initialized. The clean-up of recovery
points in each of the stores is handled by the caller

Change-Id: Ifce5e23a0d81deca16f68a0c6692cbe4ed6060cd