MB-35328 Use nodeUUID instead of hostname in StreamBegin messages
The hostname of a node can change after it is added to cluster.
Using hostname for handling KV failover from indexer can lead
to indexer getting stuck in shutdown-resart loop.

Instead of using hostname for handling KV failover, it is better
to use nodeUUID as UUID of a node is set once a node joins a
cluster and the only way to change the UUID of the node is to
remove it from the cluster and add it back

Change-Id: I6932bf5b08dbc1ea73fc6f5dd19685cd793ca5a6