MB-36386 Drop replica does not seem to drop replica
In planner.DropReplica, we return without dropping replica if the
cluster has fewer replica than requested. At this point, instead of
comparing the number of replicas in the cluster, we tried to compare
the number of partitions of a replica resulting in this bug.

Updated the code to compare the number of replicas in the system

Change-Id: I19a214656c4e580b7826c43a7a9bae8a4a119e51