MB-36582 use failover log to validate vbuuid
before using a disk snapshot, use the failover log to
validate vbuuids. this can helpful in cases where
memcached loses a vbuuid that was sent to indexer(either
due to failover or memcached crash). In case only a
vbuuid has been lost but no mutation, this will avoid
the unknown vbuuid error.

Change-Id: I34340592c9bdcce31ca4cda45f039466d3d49e64