MB-36601 Force nodeToHostMap update during initilization
The monitorKVNodes() function call is made at the end of init
phase in indexer. This method will listen to all pool change
notifications and update the nodeToHostMap() when there is a
change in KV nodes.

When indexer restarts, the monitorKVNodes() can miss a pool
change notification. If no new pool change notification arrives,
then the nodeToHostMap is not updated. Also, the logic for
periodically monitoring KV nodes did not work as cluster info cache
was not getting updated periodically. Since the cluster info cache was
never updated after restart, the nodeToHostMap is also not updated and
hence the "Erorr in book-keeping" messages

This patch will force the update of nodeToHostMap when monitorKVNodes
is called. Additionally, each call to getActiveKVNodes will update the
cluster info cache. This will make sure that nodeToHostMap remains
updated even after indexer restart

Change-Id: Ieeb8167e3f444f1b4008a072985476035024fe90