MB-36745 Remove logging when node not found in nodeToHostMap
When a KV node is failed over or reblanced out of cluster, the corresponding
entry in prjLatencyMap is deleted only when StreamBegin messages for the
relocated vb's arrive at indexer. However, nodeToHostMap is updated as soon
as rebalance is initiated.

During this period, StatsManager will keep logging the node not found in
nodeToHostMap error message. This can unnecessarily flood the logs. Hence
removing the log message

Change-Id: Ife751b38d45790527ca71a2ce12b16c5ccc8f6ca