MB-36464 Return if there is an error in updating Pools
Currently, indexer does not check if there was any error in updating
Pools structure. It later uses the entries from BucketURL map. If there
was an error in updating the Pools structure, then BucketURL map would
be empty. This would result in indexer querying the wrong endpoint during
refresh() method

In this patch, indexer does not proceed further if it sees any error while
trying to update the pools endpoint

Change-Id: I189edfd57f70e391b992f80ab312b19ea45308df