MB-36964: Fix recoverableCreateIndex in case of planner error
If planner fails to generate a solution that satisfies the
connstraints, a round robin approach is used to allow index
creation. But the layout generated by the round robin approach
is not used during commit phase.

The fix ensures that the layout generated by the round robin
approach is used during commit phase.

Change-Id: Ie108946cb75c0d81e3c0d92a5742c7e062866531