MB-36982 Update rawDataSize only if delete is successful for primary index in MOI
Currently, deletePrimaryIndex() in MOI does not check if an entry exists in
it's store or not. Due to DCP snapshot de-duplication, it is possible that
DCP sends DELETE messages for items that are not indexed. In such a case,
raw_data_size stat could go -ve.

With the fix, we try to update the raw_data_size stat only when delete is
successful

Change-Id: Id9a57dc8dff1691e5ade277af2c73d31d0ae9108