MB-36979: Fix minimum memory of new index getting created
When an index is getting created, its memory size, min memory
size etc. are need to be estimated. In case of min memory
estimated size should be non-zero so that planner can better
calculate the cost of the solution after index movement.

So, using estimated memory size as min memory size for new
index that is getting created.

Change-Id: I35076b910bb03e1daeed8698135c597c5989580d