MB-37361 Optimise feed.addInstances calls to kvdata.AddEngines
Currently, the number of times kvdata.AddEngines() is being called
from feed.addInstances() is equivalent to the number of instances
to be added.

All the instances belonging to one bucket that are sent with addInstances()
are being added to the corresponding kvdata and workers in the first call to
the corresponding kvdata.AddEngines(). The other calls are redundant.