MB-37448 [BP 6.0.4] store last vbuuid with valid mutation
store last vbuuid with a valid mutation. this will be used to
retry dcp stream request in case of rollback. this is useful
in cases where memcached can loose its vbuuid due to crash or
takeover. this mechanism can help avoid the "unknown vbuuid"
error and subsequent rollback to zero.

Change-Id: Ica833672b4c5409c973c8125085c53e4cb76d559