MB-31959 Use keyspaceId for DCP conn. name and update feeder book-keeping to refer with keyspaceId
The openFeeder() method opens an upstream feed with the bucket name.
With collections, the projector book-keeping is made at keyspace level
and not at bucket level. The Keyspace name is chosen by indexer. This
patch updates the openFeeder() method to open the upstream feed with
keyspaceId name. It also updates the feeder book-keeping to refer with
keyspaceId rather than with bucket

Change-Id: I5d7de176a1bd95233395c4753eddcd27fb1dfb88