MB-31944 New infra for rollback to zero
Any index receiving a rollback to zero will again go through
the lifecycle similar to index creation/build within indexer server.
This change is aimed at supporting the collection flush with a
stateless approach.

This also helps in regular cases where after a rollback to 0, index
scans would return no results till first big DCP snapshot is processed.
Now after a rollback to 0, index state will be changed to Create/Initial
and all scans will return a proper error message.

Support for OSO mode needs this new infra as well. If an index
get rollback to 0 and DCP moves to OSO mode, it is not possible
to continue the index in Active state.

Change-Id: I68cc1fef6d74d6d31a153201a8da167597f34020