MB-37548 Re-generate protobuf files (.pb.go) when .proto files are updated
In CMakeLists.txt, even though ADD_CUSTOM_TARGET gets executed everytime,
the COMMAND in ADD_CUSTOM_COMMAND gets executed only if the OUTPUT files
(i.e. generated files) are out-dated compared to the files the command
DEPENDS on.

Currently, the commands depends on "secondary/protobuf" directory whose
timestamp does not get updated when a .proto file is updated. This does not
re-trigger the protoc command.

This fix changes the dependency of ADD_CUSTOM_COMMAND to depend on the corresponding
.proto files rather than the "secondary/protobuf" directory. With this fix,
any changes to .proto files will trigger the generation of .pb.go files

Change-Id: I895b8d4aa0ccb23be3da716f6ec4015e034d8785