MB-36524 Update test framework to handle partitioned indexes
The current test framework for getting stats does not account for
partitioned indexes. This is causing the TestPartitionDistributionWithReplica
test to fail with mismatched scan requests.

This patch updates the test framework to accumulate the stats for the
same index from multiple nodes to accommodate partitioned index. New methods
are added to get per partitioned indexes. The scan validation is updated
to validate the num_requests per partition.

Change-Id: I3a2c6335e1e9143e7c8e29cd863d7a7135b64405