MB-31955 Add support for DCP system events at projector
The DCP_SYSTEMS_EVENTS pertaining to CollectionCreate,
CollectionDrop, CollectionFlush, ScopeCreate, ScopeDrop,
CollectionChanged are propagated to down-stream with this patch

Five new fields are introduced in DcpEvent structure. These are:
(a) ManifestUID (b) ScopeID (c) CollectionID (d) EventType (e) MaxTTL.

ManifestUID and ScopeID are stored as slice of bytes.

CollectionID is stored as Uint32 to avoid multiple conversions (It is
easy to decode collectionID as Uint32 from mutation key. Also, filtering
at KVData is relatively simple with Uint32 values when compared to slice of
bytes).

EventType represents one of the 6 system events. MaxTTL is the maximum ttl
value of the collection

The key, oldKey, partnKey in KeyVersions are mis-interpreted to represent
manifestUID, scopeID and collectionID while sending data to indexer.

Change-Id: I023d03b36c081611d1006790c1e756c3e3ebab3a