MB-37425 Return latest version from cluster info cache
When queried for cluster version, cluster info cache returns
INDEXER_55_VERSION as default even if cluster is on later
version of the server (say 7.0). This could break some QE tests which
depend on later versions of indexer.

This patch, updates the code to return INDEXER_CUR_VERSION by
default and returns INDEXER_70_VERSION if the major version
number if 7. Also, INDEXER_CUR_VERSION is updated to refer to
INDEXER_70_VERSION

Change-Id: If154a0aaba4610497c09d14dfd056f3e3fd13536