MB-31109: Make ip:port binding on projector and indexer more lenient
If the cluster is configured to use ipv4, allow GSI processes
to come up successfully even if they cannot bind to ipv6:port.

Similarly, if the cluster is configured to use ipv6, allow GSI
processes to come up successfully even if they cannot bind to
ipv4:port.

Note that the GSI clients will use the node names from cluster
info cache. And the cluster configuration change with respect
to ipv4/ipv6 protocol stack is not supported if the cluser
node names are based on the ip addresses.

Change-Id: Iadb546e60ef32a3edd8ce3e5e41be8d1e721b443