MB-31955 Support filtering of mutations at worker based on collectionID
When DCP stream is opened for entire bucket, mutations beloning to all
collections will be streamed by KV. However, not all mutations need to
be evaluated. Only those mutations that belong to the collections for
which instances are added at worker will be evaluated. For all other
mutations, an UpdateSeqno message will be propagated to indexer for
book-keeping updates

Change-Id: I1cc0b05af223c1de462171bb1e5d0457d07caa75