MB-38029: [BP 6.5.1] Set immutable flag to false when index is partitioned and partial
BP of MB-37737 to 6.5.1

Immutable flag is always set to true if an index is partitioned index. This
breaks upsertDeletes for an index that is partial and partitioned index. The
fix sets the immutable flag to true when the index is partial overriding
partitioned index.

The fix also changes flusher behavior to avoid un-necessary slice Deletes.
The new design is:
  UpsertDeletion - if partition key is available, delete from the partition
  the mutation belongs to.
  Delete mutation - Delete from all partitions as we will not have partition
  key in case of a delete mutation

The fix also includes a change in projector to always evaluate partition key
irrespective of whether WHERE evaluates to true or false.

Also, this patch does not attempt to fix metadata of an existing partitioned +
partial index, however sets the right Imuutable value in index definition in
mutation manager.

Change-Id: I16bdecaeff4be021f711ec6c74676a49ad776442