MB-38190 Fix channel length stat calculations in projector
BP of MB-38151 to 6.5.1

Currently, the channel length is calculated inside control loop of
kvdata, worker, endpoint components. In case where down stream is
stuck, the channel length contains the last updated value and not
the actual value of channel length at the time of stat collection.

This patch updates the stats object to contain reference to channel
so that the channel length can be calculated asyncronously

Change-Id: I81a79dee94150bff985865413f951997d9913465