MB-33849 Split control and data processing to different go-routines in worker
The control and data processing is split into two go-routines
with this patch. genServer() handles the control path and run()
handles the datapath. Between the two go-routines, engines, endpoints
and vbuckets maps are shared.

The engines, endpoints maps are read-only in run() and are updated
in genServer() routine. The vbuckets map is read-only in genServer()
and is updated in run() routine.

As there are two separate go-routines, there is no ordering in the
way the routines can terminate. To make sure both the go-routines
exit gracefully, the following channels have been introduced:

a. genServerFinCh - Controls the execution of genServer() routine.
genServer() will exit when this channel is closed

b. runFinCh - Controls the execution of run() routine. run() will
exit when this channel is closed

c. genServerStopCh - This channel is used to block all incoming
requests reaching worker. Incoming requests will error out with
ErrorClosed when this channel is closed

The semantics followed are:

a. When run() exits, then close genServerFinCh so that genServer()
terminates (In case run() exists first, closing genServerFinCh ensures
that genServer() will eventually exit). Also, StreamEnd messages
are published to downstream to facilitate vbucket book-keeping
clean-up

b. If genServer() exits, then
    1. Close genServerStopCh to block all incoming requests to this
       kvdata instance
    2. Close runFinCh so that run() terminates and data processing
       is stopped for this worker instance

Change-Id: Ibb90e27d837767ae05646e10127d551225474141