MB-38386: Fix size estimation of an empty index in index planner.
Backport MB-38194 to 6.6.0

If the data size and memory usage for an index are zero (actual
or otherwise), the index planner treats that index as NoUsageInfo
index. It means that the usage info for that index is not available
and the sizing estimation needs to be done for that index.

For MOI index created on a bucket which never had any mutations,
the data size and memory usage can be zero (even if the index state
is INDEX_STATE_ACTIVE).

Size estimation assumes that the indexes without usage info will
take up all the free memory from all indexers. This leads to
wrong sizing estimation for empty index. This can further lead
to planner yielding non-uniform index distribution.

As a fix, size of an empty index - with state as INDEX_STATE_ACTIVE,
will be estimated to a small constant value - which is a realistic
size estimation for an empty index.

Change-Id: I61c4ff525234d8b654ec2be57e9fbc1de0b8418d
(cherry picked from commit d2be7a0ac557d7bfcb74b68a396010338db7e8d9)