MB-31938 use bucket as keyspaceId in timestamp
It is better for the actual timestamp to have bucket as the keyspaceId
a. the index snapshot carries the timestamp in its metadata. the index
gets build in INIT_STREAM and then moves to MAINT_STREAM. The scans
happen in MAINT_STREAM. It is easy to have same keyspaceId during all
these states.

b. upgrade is easier with this approach

Change-Id: I7659b3d230109cd4bbbc8712af00328c22ae8bde