MB-31938 add collectionAware flag to MTR request
It is not sufficient to use feed version to decide whether to make helo
collections request to memcached as it will error out if the cluster is
in mixed mode.

Indexer will specify a collectionAware flag in MTR to let projector know
if the cluster is out of mixed mode and the DCP can be requested to
enable collections.

Change-Id: Ied08d9a055b7cb2f8fa3a025280a92cae581a9c3