MB-39554: Planner should always run estimation for unbuilt indexes
Index planner runs index size estimation based on the values of
memory size and data size. If both of these values are zero,
size estimation is run. In some cases, data size of the index
can be non-zero even if the index is unbuilt.

So, forcing size estimation based on index state.

Change-Id: I38977169cf6b92c38f9f401143475fe605ec49ff