MB-39148 Construct slice cmdCh to contain reference of indexMutation to reduce idle time RSS
The size of indexMutation struct is 64 bytes. By default, 50,000 entries
of indexMutation will be allocated for cmdCh across all slices. This
comes to 50000*64 = ~3.05MB per index. With 1000 indexes, the cmdCh in
each slice would occupy around 3G.

In this patch, instead of storing the indexMutation struct in cmdCh, we
store a reference to indexMutation struct. This means that cmdCh across
all slices would only require 50000*8=~390k.

Experimentally, the idle indexer RSS has come-down by half (from 6G to
3G) with 1000 indexes and MOI storage. The size of cmdCh has comedown
from 3G to 250M.

Did not notice much of a change in index build times.

Change-Id: I73f4d1b72e279ae0590755dfdc8613d6002f7ffe