MB-39148 Make the default size of sliceBufSize proportional to numWriters
Currently, the default size of sliceBufSize is set at 50000. This value is
shared across all writers in a slice. This can become a memory overhead
when there are large number of idle indexes.

Experimentally, noticed that the time spent in wait on cmdCh in slice
before and after reducing the sliceBufSize was around 5.8 sec (for 20M
mutations, MOI storage) and the index build time was not impacted.

Default numWriters are configured to runtime.GOMAXPROCS. So, the default
value of sliceBufSize is made proportional to runtime.GOMAXPROCS

Change-Id: I2495438f41a9f722acb9f9e12435fcf9ee956913