MB-31951 Use collection aware Seqnos for consistent scans
A global cluster version is maintained which keeps a track
of the cluster version when a connection to KV is made.

The GetSeqnos() API uses this global cluster version when
trying to decide between calling CollectionSeqnos() (or)
BucketSeqnos(). Since all indexes from version >= 7.0
will be defined on collections, it would call CollectionSeqnos()
for a upgraded cluster.

A go-routine asyncronously monitors the "/poolsStreaming" endpoint
and updates the global cluster version when there is a change.
Also, the cluster version is updated when establishing connections
with KV nodes

Note: While service change notifier can be used to watch cluster
version changes, when there are multiple subscribers to services
change notifier, some of them may take long time to process the
notification in some cases thereby delaying the notification by
more than 5 seconds. To avoid this scenario, this patch directly
monitors the "poolsStreaming" endpoint

Change-Id: I7346fdcfbd2e044e4dd9f1c113b5e4f1c4292c6f