MB-39658 Do not broadcast list of indexes if there is no change in list of indexes or buckets
The protocol is as follows:
a. When a watcher starts, it sends a message to indexer and get's
the client stats and caches them
b. From then on, when ever watcher gets stats, it checks if the map
of index names is empty or not
c. If the map is empty, then it will retrieve the name of indexes
from the cached client stats
d. Otherwise, watcher will update the cached client stats with
the index names present in the broadcasted message

Change-Id: I57d946e0fbd06edd8d161b855d4ae18d64e9b9b9