MB-39698 Optimize the computation of equivalent indexes
The current logic used to compute equivalents has a complexity of
O(n * n). This doesn't scale well with increase in number of indexes.

With the current patch, the list of indexes are traversed first and
then segregated according to bucket, scope and collection. Only
all indexes belonging to a collection are checked for equivalence.

Change-Id: Ied2d802d9a5243eed46e5eb332194aad08a6a3d1