MB-39148 Reduce the default size of encodeBuf
When allowLargeKeys is set to true, the encode buf with
each slice is re-sized based on the incoming key size.
When there are large number of indexes in the cluster,
encode buffers significantly contribute to idle time
memory utilisation. Reducing the default size of encode
buffers would reduce the idle time memory utilisation of
indexer process

Experimentally, with this patch, it was observed that the
time spent in resizing encode buf is relatively less.

E.g., when building an index with monotonically increasing
key sizes from 3kb to 4kb, in steps of 10, the total time
spent in re-sizing encode buf is 400us during an index
build duration of 75 sec.

Similarly for random key sizes in the range of 3kb-4kb, the
total time spent in resizing is 163us during an index build
duration of 78 sec

Change-Id: Iafee1a1dc0571f143d8998eb829eefe47eb17f4c