MB-40989 Use separate cluster info client for lifecycle manager
It is not a good idea to share cluster info client between manager
and lifecycle manager as all methods in request handler are serviced
using cluster info client of manager. So, if lifecycle manager has to
refresh its cache while indexer is servicing a request like
getIndexStatus, lifecycle manager has to wait for getIndexStatus
to finish. Depending on the number of indexes in the system, this can
take long time.

Change-Id: Ic68ec856d4636cd132256ea8dc873b713f3b647a