MB-41179 Increase finding watcher duration to 25 sec during retry
With large number of indexes, watcher is going to take more time
to sync the index definitions and metadata from peers. When there
is a DDL during this time, it can fail because watcher is still
syncing the data and not ready for serving the DDL. In such case,
the DDL can fail with indexer not available error.

This patch addresses this issue by increasing the watcher retry
duration

Change-Id: Ib595006e72704951e2127728a843ca3ad40f7b30