MB-31952 Use RBAC cache to avoid thrashing ns_server
ns_server being the main consmer of getIndexStats, it will
call this endpoint with full cluster access. Therefore, it
becomes sub-optimal to query ns_server for RBAC for each
collection.

As RBAC is hierarchical i.e. a user with bucket level access
will have access to all scopes and collections in the bucket
and similarly a user with scope level access will have access
to all collections in the scope, the validation is also made
hierarchical.

If a user has bucket level access, then indexer would not query
for scope and collection level access to ns_server. This way, the
number of RBAC requests to ns_server will come down

Change-Id: I55f5ecef0c11c755c6a9a179e86da17985867b88