MB-41062: Add new Cluster Info Client for Request Handler.
getIndexStatus was wating for around 20 Seconds in the worst case
to acquire lock for cluster info so creating a new cluster info
client i.e. reqcic for RequestHandlers.

Mutex Profiling showed goroutines contending for lock as below
* getKeyspaceForCleanup
* watchClusterChanges
* getIndexStatus
When occured in the above order Lock being in between RLocks
is causing delays. When moved to new ClusterInfoClient we are not
observing the 20Sec Delays occasionally.

As Querying isEphemeral without RLock is not advisable and can lead
to panics adding RLock when accessing cinfo.

Change-Id: Iab7f3d435b9437771e283cf11d5963695dc5bc57