MB-41288 Return all storage stats when queried from REST endpoint
For REST endpoint, we need to return all the storage statistics
when queried for. In the REST handler for storage stats, the consumerFilter
will be set to "AllStatsFilter" when no consumerFilter is specified and
therefore will return all stats for this plasma instance irrespective of
StatsLoggingEnabled flag

For stats logging, consumerFilter will take the default uninitialized
value of "0" (In storage_manager, it is uninitialized). Therefore, if
StatsLoggingEnabled is false for those instances, then we do not get
plasma stats

Change-Id: I654ae830f9ff798f7b0ff57493ce7b98f0695aca