MB-42108 [BP 6.6.1] skip cleanup of valid partition even if tombstone exists
If the real instance is holding a valid partition, it should not be
cleaned up on warmup, even if a tombstone exists. The tombstone could be
present due to bug in cleanup mechanism.

This change allows for an existing cluster, which has run into the bad
tombstone issue, to upgrade safely using failover/recovery mechanism.

Backport of MB-42160

Change-Id: I29359e42921ef4aae7b692acfb5fb4f31ca03b5e