MB-41529 Do not delete indexes from other buckets on collection drop
On collection drop, there is no check to validate the bucket to which
the collection belongs to. As collections from different buckets can
share the same collectionID, on a collection drop, indexer is dropping
all indexes across all collections.

With this fix, indexer would validate the bucket to which the collection
that is dropped belongs to and then cleans the indexes accordingly

Change-Id: I5f14cd8962d459a52ce7ec9d1a5c85bda84b0200