MB-32052 Check if debug logging is enabled before debug logs
The logging.Debug statements take slice.IsDirty() as argument.
This means that slide.IsDirty will be called irrespective of
whether debug logging is enabled or not. Calling this method
is un-necessary as it invokes the handleCommandsWorker go-routine
and this incurs some CPU.

This patch adds a log level validation before calling the debug logs

Change-Id: Ia99d66b9dedbd63291d276a2b6771a841d84986c