MB-43280 - [BP] Reduce time taken for listReplicaCount
Back porting to 6.6.x of MB-42063 - Reduce time taken for listReplicaCount

As the number of create/drop tokens in Metakv are less in number when
compared to number of indexes in node, while fetching tokens for calculting
ReplicaCount it should be less expensive to fetch all tokens and update
the count than checking for a token for every index.

Time taken for getLocalReplicaCount function is improved from 2sec to 5ms
on average.

Change-Id: I6731ce2ea6fdebf70f0b6f256e5da3aa299a1cad