MB-39161 Do not retry to fetch clusterinfocache when cleaning deferred indexes on keyspace drop
In lifecycle manager, reaching handleCleanupDeferIndexFromKeyspace code
path means that the cluster info cache has either lost information about
the keyspace (or) new information about the keyspace is updated. Therefore,
force fetch cluster info cache incase keyspace ID is NIL is not required.

When the keyspace is dropped, a notification will be sent to the cluster
info client of both lifecycle manager and manager components. If manager has
updated it's client, and if lifecycle manager does not update it's client
for more than 5 sec, then client will force fetch cluster info cache. This
means that the next iteration of monitorKeyspace will identify the indexes
that have to be dropped and will send a message to drop them

Change-Id: I29c9e3b4f911eb46fe107ef5bee3dfcb7ec6b933