MB-32052 Check is flush is active before evaluating the dirtyness of a slice
When there are large number of indexes in MAINT_STREAM and if a
mutation comes for a collection, then the timestamp returned by
flusher will be greater than the snapshot timestamp of all the
instances in the MAINT_STREAM

However, only indexes belonging to the collection for which mutation
has arrived will be dirty. Evaluating slice.IsDirty() for all other
instances is un-necessary. Also, slice.IsDirty() is a costly operation
as it has to wake-up all slice writers every 10ms (in-memory snapshot
interval).

This patch addresses the problem by checking if the flush is active
or not before checking the dirtyness of a slice. Flush is considered
active when mutations are being written into cmdCh's of slice writers.
Flush is considered "Done" when a snapshot is generated (because, at
the time of snapshot generation, the cmdCh would be empty).

This check will avoid calling slice.IsDirty() for all instances in the
stream. Rather, it would check the dirtyness only for those instances
that processed mutations

Change-Id: I1ac3f5eacf3b9edb56192d5972aaf0fb4ee13a88