MB-42890 Preserve the mapping between keyspaceId and reqTs,rollTs,actTs
In topic response, the keyspaceIdsList is built based on the ordering
of keyspaceId's in actTss. It is not guaranteed that golang maintains
the same ordering of keyspaceId's in reqTss and rollTss.

E.g., actTss can store keyspace ID's "a","b", "c" in the same order
but reqTss can store it in "b", "a", "c" and rollTss in the order
"c", "b", "a"

This can lead to inconsistent mapping between keyspaceId and reqTs,
rollTs in topic response i.e. the reqTs of one keyspaceId can be
processed by different keyspace. This patch fixes the issue by
preserving the order.

Additionally, the feed's timestamp book-keeping code is built on
some invariants which should never be violated. Added some checks
and logging information around those invariants for easy debugging
if some thing goes wrong due to a bug in other parts of the code

Change-Id: I8419b39acb29e90f3ae1945056e2f5d8a7bd6825