MB-43104 Fix kvutility test code to read and close resp.Body
When client.Do() returns a success response, the caller is required to both read to
the end of the response and close the response body, else the TCP connection is not
freed, which may eventually result in the test framework failing with error

  ERROR: TOO MANY ACTIVE CONNS 251

from queryport/client/conn_pool_test.go

2020-12-08 Amended to address Varun's review comments where he noticed a bucket also
not being closed. I found another case of unclosed bucket as well. Fixed those plus
changed all bucket closes to "defer b.Close()" right after the error check so they
will be closed even if the function hits another error (as many of them can).

Change-Id: I5c7fd1bfbd9ed45ab0460a6f91bdd9f6afe342d9