MB-43194 Fix TCP connection leaks in querycmd/docmd.go HandleCommand()
case "config" in querycmd/docmd.go HandleCommand() was leaking two TCP connections
every time and a third one if there was an error during processing of its response.
I fixed all of these to use "defer" to close the response body on successful
responses, and also to aways try to read to the end of the response body. Both of
these actions are required by Go's client.Do() method to avoid leaking the TCP
connection it uses.

Change-Id: Ie275ef7dbd7d14823eb14bf8d2ce80e2c12a47fd