MB-43263 Delete and add default bucket in TestDeleteBucketWhileInitialIndexBuild
Depending on the memory usage of default bucket, decreasing
the bucket memory quota from 1024MB to 256MB after
flush operation is taking more than 30 seconds in some cases.

This is causing the error: "RAM quota cannot be set below current usage"
and the new RAM quota is not getting applied.

Hence, deleting and re-creating the bucket to avoid test failures

Change-Id: I8b6abe5ee87d8f63f9900a70332308d1cab16bdc