MB-43226 Use lock-less approach to get dcp_getseqnos timing stat
BP of MB-43279 to 6.6.2

In the current code, getting the dcp_getseqnos timing stat requires
the dcp_buckets_seqnos.rw read lock. Acquiring this lock can take time
when dcp_seqno.go is busy establishing connections with KV.

The stats endpoint used by indexer requires getting the seqnos timing
stat. In some cases, it was observed that the call to stats endpoint
with async=true timed out (when the timeout is 10 sec) due to this
lock dependency.

This patch addresses the issue by removing the lock dependency using
Copy-On-Write approach for vbSeqnosReader map

Change-Id: I50519b3e0f82a67927b40a079b43258e2c6cd7d3