MB-36789 Rename "Replicating" state to "Moving"
Index moves (from alter and rebalance) had confusing state string "Replicating",
which sounds like a master-to-replica build process. Changed it to "Moving".
The check in rebalancer.go needs to check for both "Moving" and "Replicating" as
we support nodes of different releases in the same cluster, so pre-7.0 nodes
will still report "Replicating".

Change-Id: I5dec6340f5f060e5db1422eeb77521e797406c0d