MB-43157 More specific error code from GSI client for index not found
Fix is to wrap the GSI-specific error inside the data service error
returned by errors.NewCbIndexNotFoundError() (since we cannot pass
the GSI-specific error message, which includes the index ID that
we'd want for troubleshooting, into that function – it hard-codes
"Index Not Found" as the message).

This is accomplished by changing
  return nil, err
to
  return nil, errors.NewCbIndexNotFoundError(err)
in IndexById() and IndexByName() (secondary_index.go).

Change-Id: I155798af0f132b1c3cfc00d7ff6c2af9e8693e7a