MB-40554 Notify waiters on orphan definition removal
recoverableCreateIndex will wait for a notification from
updateTopology when a non-deferred index is created. However,
if the index is dropped when the build is in progress and the
index has only one instance, then updateTopology would consider
it as orphan definition and will remote it from it's list of
indices.

Because of this, the caller of recoverableCreateIndex will wait
forever. With this patch, when an orphan definition is being
cleaned up in the metadata provider, then all the waiters will
be notified.

Change-Id: Idb9177ff62a5ea0d44e42b536a81bde872bfc152