MB-36987 Use addressFamily from pools/default to validate IPV6 nodes
The "isIPv6" option from /pools endpoint will be depricated.
The suggestion from ns_server team is to use "addressFamily" of each
node to determine if the node uses ipv4 or ipv6 protocol.

This patch updates the code to retrieve "addressFamily" of each of the
node and validates cluster info cache based on the addressFamily (The
isIPV6 option is kept as it is to support mixed more cluster scenarios)

Change-Id: I5ecb4944cbca93136a2a47e828d4d30ea1fe8135