MB-39475 cbindex: -batch_build - Avoid scheduling indexes for build
With http://review.couchbase.org/c/indexing/+/140676, indexer takes
care of building indexes for upto 10 collections at a time. If the
number of collections being built simultaneously at a time exceed
10, then indexer will now add the build commands to queue and return
an error that the index will be built later.

In this patch, moving the existing "batch_build" option to
"batch_build_schedule" and keeping the behavior of "batch_build"
similar to that of "batch_process"

Change-Id: Iaa72fa6ffcc2afe49011e550fadc43af8124f083