MB-42194 Enable firstSnap optimisation for deletes in MOI
This will speed up initial index build performance when there
are large number of partial indexes being built at the same time.
With this optimisation, MOI doesnot have to process upsertDelete
messages from projector when the mutations belong to first snapshot

For updates, the firstSnap optimisation is inherently present
in the code. So, it does not have to be handled explicitly

Change-Id: I403fc32759e0dba9c675282f57002ecbfa00ac7f