MB-43259 Stop mutation scatter before draining the mutch in kvdata
When feed decides to clean-up the keyspace, it spawns a go-routine
to drain the mutation channel in kvdata. This is done to ensure that
the StreamEnd messages published by dcp_feed.go routines will be
consumed if the kvdata mutch is full.

When this go-routine is spawned in cleanupKeyspace, there will be
two consumers of kvdata's mutch. In such a case, it can be possible
that important control messages like Snapshot marker message gets
drained while the mutations belonging to that snapshot are processed
at mutch in kvdata. This breaks the DCP logic as indexer always
expects a snapshot marker message before a mutation from that snapshot.

The issue is fixed in this patch by stopping the runScatter method
before processing any mutations. Feed sets the stopScatter variable
to "1" which will break the runScatter loop.

One other way to fix this issue is to just call kvdata.Close() instead
of setting the stopScatter variable. That approach will not work in
some cases after MB-33849 is fixed i.e. when control and data messages
are processed in separate go-routines. After the go-routine processing
control messages is closed, there can exist some time for the go-routine
processing data messages to get closed. This period of quiscence can
break the ordering of messages to indexer. To avoid such scenario,
feed closes the data path first so that no new mutations can be
forwarded down-stream after the current mutation.  All the remaining
mutations will be drained the go-routine spawned by feed

Change-Id: Ia036802624b76cc7cdfccddcc86b339fb63f782c