MB-42743 Avoid parsing all instances in findIndexInstance for scans
With 1000's of indexes on a node, iterating over list of all indexes
will increase the scan init latency. This can be avoided by maintaining
a indexDefnMap which keeps a track of all instances belonging to an
index definition. That way, ScanCoordinator::findIndexInstance will
only have to iterate the instances of interest

Change-Id: I93bf9d229eb31bdcf6ac77e3b774e158d9dc221a