MB-42623 Fix index scan special case for descending sort
Scans of partitioned indexes that request the primary key
plus all GSI keys in the same order they are defined in the
index, where at least one of those keys is indexed in
descending order, were returning unsorted data because
GsiClient code analyzeProjection() (in
queryport/client/scatter.go) was making an incorrect
assumption that if sorted data is required, there will
always be a projection list, but in this case that is not
true, so it was not setting up the scan request's projDesc
slice for this case.

The fix is to copy the IndexDefn.Desc slice to the scan
request (RequestBroker.projDesc). Without this the code
will think the requested ordering does not match the index
ordering and will not bother to sort the rows.

Change-Id: I4064e0755356b9fb615908ac99019e12823ff45d