MB-43994: Fix DCP feed names
It is not required for DCP feed names to contain bucket, scope and
collection names. With large names of these keyspaces, the DCP feed
key size limit can reach and feed request will be rejected by DCP
producer.

Projector does not maintain the feeds based on feed name, but it
maintains the feeder object using keyspace as key.

Projector already guarantees the uniqueness of the feed names by
adding uuid to it.

Change-Id: I6fe1a6b74b7b082d44c470d17585e8a6a53ca0d5