MB-44049 Do not destroy an already destroyed snapshot
When there is an index delete, storage manger would send
a notitication to scan coordinator. Previously, the
lastSnapshot map used to delete the snapshot of the dropped
index while holding the mu.Lock() in listenSnapshot go-routine.

After avoiding the centralised lock in listenSnapshot(), the
deletion of instance is handled when indexer sends the updated
instance map to scan-coordinator. When this message is received,
the snapshot is also destroyed. This leads to the snapshot being
closed twice - Once when delete notification arrives and once
when instance map is updated. This patch will make sure that
the snapshot get's deleted only once

Change-Id: Ic32432ecf9d006d438ec0a21577a846a6ab76fed